// src/api/generated/models/projectHistoryResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { ProjectHistoryResourceTitlePt } from './projectHistoryResourceTitlePt';
import type { ProjectHistoryResourceTitleEn } from './projectHistoryResourceTitleEn';
import type { ProjectHistoryResourceDescriptionPt } from './projectHistoryResourceDescriptionPt';
import type { ProjectHistoryResourceDescriptionEn } from './projectHistoryResourceDescriptionEn';
import type { ProjectHistoryResourceTranslationStatusLabel } from './projectHistoryResourceTranslationStatusLabel';
import type { ProjectHistoryResourceProdUrl } from './projectHistoryResourceProdUrl';
import type { ProjectHistoryResourceGithubUrl } from './projectHistoryResourceGithubUrl';
import type { DevProfileResource } from './devProfileResource';
import type { LanguageResource } from './languageResource';
import type { ProjectHistoryResourceGalleryItem } from './projectHistoryResourceGalleryItem';
import type { ProjectHistoryResourceCreatedAt } from './projectHistoryResourceCreatedAt';
import type { ProjectHistoryResourceUpdatedAt } from './projectHistoryResourceUpdatedAt';

export interface ProjectHistoryResource {
  id: string;
  title: string;
  title_pt: ProjectHistoryResourceTitlePt;
  title_en: ProjectHistoryResourceTitleEn;
  description: string;
  description_pt: ProjectHistoryResourceDescriptionPt;
  description_en: ProjectHistoryResourceDescriptionEn;
  translation_status: string;
  translation_status_label: ProjectHistoryResourceTranslationStatusLabel;
  prod_url: ProjectHistoryResourceProdUrl;
  github_url: ProjectHistoryResourceGithubUrl;
  dev_profile_id: string;
  dev_profile?: DevProfileResource;
  languages?: LanguageResource[];
  gallery: ProjectHistoryResourceGalleryItem[];
  created_at: ProjectHistoryResourceCreatedAt;
  updated_at: ProjectHistoryResourceUpdatedAt;
}